move netlink funcs and create stubs
Projects that vendor types from ocicni and also must support darwin and
windows fail to build due to the netlink package.  To avoid this, move
the funcs in question into _linux and create stubs in _unsupported.

No code changes to the functions were made.
