+++ pkg/ocicni/util_unsupported.go
@@ -3,17 +3,32 @@
 package ocicni
 
 import (
-	"fmt"
+	"errors"
 	"net"
 )
 
 type nsManager struct {
 }
 
+var errUnsupportedPlatform = errors.New("unsupported platform")
+
 func (nsm *nsManager) init() error {
 	return nil
 }
 
 func getContainerDetails(nsm *nsManager, netnsPath, interfaceName, addrType string) (*net.IPNet, *net.HardwareAddr, error) {
-	return nil, nil, fmt.Errorf("not supported yet")
+	return nil, nil, errUnsupportedPlatform
+}
+
+func tearDownLoopback(netns string) error {
+	return errUnsupportedPlatform
+}
+
+func bringUpLoopback(netns string) error {
+	return errUnsupportedPlatform
+}
+
+func checkLoopback(netns string) error {
+	return errUnsupportedPlatform
+
 }
